Money types: agriculture BigDecimal, education whole-number fees
Currency must never be a binary float. Two services were off-standard
(business/finance already use BigDecimal(19,2)):

- agriculture: AgricultureExpense/Income.amount + AgricultureBaseDTO.amount
  were Float -> BigDecimal(19,2). Flyway V3 MODIFYs both columns to DECIMAL(19,2).
  Entity + DTO both BigDecimal so ModelMapper stays type-safe.

- education fees are whole-currency-unit BY DESIGN (discount.amount, the whole
  fee_collection ledger and vehicle fare are all INT). Only grade.fee/student.fee
  were the anomaly (Float) -> Integer, matching the rest of the system. FeeService
  math is now integer, a percentage discount rounded to the nearest unit. Flyway
  V3 MODIFYs grade.fee/student.fee to INT.

+++ microservices/education-service/src/main/java/com/myplus/education/dto/EducationDTOs.java
@@ -47,7 +47,7 @@ public class EducationDTOs {
         private LocalTime timeTo;
         private String status;
         private Long schoolId;
-        private Float fee;
+        private Integer fee;   // whole-currency-unit (matches Grade.fee)
         private Long room;
         private LocalDateTime dated;
         private LocalDateTime updated;
@@ -96,7 +96,7 @@ public class EducationDTOs {
         private Long discountId;
         private Integer nd;
         private String di;
-        private Float fee;
+        private Integer fee;   // whole-currency-unit (matches Student.fee)
         private Integer dueDay;
         private Integer vf;
         private String pob;
